Refactored routes to their own modules. Fixed AI chat to use MCP
